Revised interface, promise returns requestId, tests broken
Realized that for a DB backend, wouldn't even have the requestId right away, which is needed for a pro-active (polling)
interface like would be needed for an HTTP API Interface. So we need to return a promsie to fulfill with the request
id and then provide both a pro-active and a reactive (promise-based) interface for reacting to the request's state
changes.

Tests are currently broken because I haven't converted most of them over
to the new interface.